requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
python-dotenv==1.0.0
supabase==1.0.3
requests==2.31.0
Flask-CORS==4.0.0
boto3==1.26.137
botocore==1.29.137 
Pillow==10.1.0